# Add 'combinations' and 'in-combinations'