New functions for `racket/list`, to accompany `*permutations`.

Besides converting the input list to a vector, `in-combinations` runs in constant space. For time:

```
(define x* (range 25))

(let ([c* (time (in-combinations x*))])
  (time (for ([x c*]) (void))))
;; cpu time: 1 real time: 1 gc time: 0
;; cpu time: 14581 real time: 15424 gc time: 777

(let ([c* (time (in-combinations x* 5))])
      (time (for ([x c*]) (void))))
;; cpu time: 0 real time: 0 gc time: 0
;; cpu time: 2 real time: 2 gc time: 0
```

If you'd rather the implementation was split across 2 functions `(in-combinations n k)` and `(in-powerset n)` I'll happily do that, but I think the name `combinations` is more intuitive. (If there's any way to redirect searches for "powerset" to return "combinations", we should add that.)