@stamourv 2016-01-07
Re search terms: Look up `index` from `scribble/base`.