@bennn 2016-01-07
Thanks for the feedback, this was really helpful.
(I'd missed `bitwise-bit-set?` earlier; also I didn't realize that `for/fold` is about as fast as `let loop ...`)

Running time is slightly faster now. Here's the average (and standard deviation) for making combinations for `(range 20)`. (Average of 30 runs.)
- before: 150.77ms, (2.77ms)
- after: 143ms, (2.65ms)