@stamourv 2016-01-07
Much better. LGTM.
Also, what happened to Robby's comment? Looks like github ate it?