@rfindler 2016-01-07
(I was just saying that two calls to shuffle in the examples might be nice.
Since you're in the neighborhood.)