@bennn 2016-01-07
@rfindler I added a call to shuffle, not sure if this is what you meant.
https://github.com/racket/racket/commit/80767a84e970dfe695cc859bd8e1489103a61396#diff-a52ea0d2b3f1810e2018c2a43bdf48beR1268