@rfindler 2016-01-08
No, I was talking about the docs for shuffle. The top line of the link
there. Don't worry about it.

Robby

On Thu, Jan 7, 2016 at 5:24 PM, Benjamin Greenman <notifications@github.com>
wrote:

> @rfindler <https://github.com/rfindler> I added a call to shuffle, not
> sure if this is what you meant.
> 80767a8#diff-a52ea0d2b3f1810e2018c2a43bdf48beR1268
> <https://github.com/racket/racket/commit/80767a84e970dfe695cc859bd8e1489103a61396#diff-a52ea0d2b3f1810e2018c2a43bdf48beR1268>
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/racket/racket/pull/1175#issuecomment-169839150>.
>