@bennn 2016-01-08
You just want more shuffle examples?