@rfindler 2016-01-08
I shouldn't have brought it up. Sorry for the noise.

On Thu, Jan 7, 2016 at 7:04 PM, Benjamin Greenman <notifications@github.com>
wrote:

> You just want more shuffle examples?
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/racket/racket/pull/1175#issuecomment-169856025>.
>