@stamourv 2016-01-08
@bennn: The answer is yes. :)
I agree that this would be an improvement.