@bennn 2016-01-08
Rebased into a single commit. Ready to merge?