@LeifAndersen 2016-01-08
Merged, thanks.